Some coverity defect fixes

+++ src/Etterna/Screen/Options/ScreenOptions.cpp
@@ -669,7 +669,7 @@ ScreenOptions::PositionRows(bool bTween)
 		if (i < first_start)
 			fPos = -0.5f;
 		else if (i >= first_end && i < second_start)
-			fPos = (((int)NUM_ROWS_SHOWN) / 2) - 0.5f;
+			fPos = (static_cast<int>(static_cast<int>(NUM_ROWS_SHOWN) / 2)) - 0.5f;
 		else if (i >= second_end)
 			fPos = ((int)NUM_ROWS_SHOWN) - 0.5f;
